Moving new unit tests to new test_trampoline_shared_ptr_return.cpp,py. In .cpp, moving structs into anonymous namespace, for compatibility with the smart_holder branch.
